+++ base/__init__.py
@@ -1,2 +1,2 @@
-VERSION = [1, 13, 3]
+VERSION = [1, 13, 4]
 __VERSION__ = '.'.join(map(str, VERSION))